+++ projects/ontimize-web-ngx/src/lib/components/o-service-base-component.class.ts
@@ -382,13 +382,17 @@ export abstract class AbstractOServiceBaseComponent<T extends AbstractComponentS
       this.queryArguments = this.getQueryArguments(filter, ovrrArgs);
 
       if (this.abortQuery.value) {
-        // Update pagination info
         this.state.queryRecordOffset = 0;
         this.state.totalQueryRecordsNumber = 0;
-        // Set empty data (order is important)
-        this.setData([]);
+        this.setData([], []);
+        /**  this.cd.detectChanges() is used to update loadingSubject value (this.loadingSubject.next(true); in line 377)
+         *  before using the next line and so update the oTableExpandedFooter directive and display the message
+         * that there are no results when the query is aborted*/
+        this.cd.detectChanges();
+        this.loadingSubject.next(false);
         return;
       }
+
       this.querySubscription = (this.dataService[queryMethodName].apply(this.dataService, this.queryArguments) as Observable<ServiceResponse>)
         .subscribe((res: ServiceResponse) => {
           let data;
@@ -404,6 +408,7 @@ export abstract class AbstractOServiceBaseComponent<T extends AbstractComponentS
               this.updatePaginationInfo(res);
             }
           }
+
           this.setData(data, this.sqlTypes, (ovrrArgs && ovrrArgs.replace));
           this.loadingSubject.next(false);
         }, err => {