+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * **o-table**:
   * Fixed bug that incorrectly set the default configuration ([1808d599](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/1808d599)) Closes [#1006](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1006)
   * Fixed bug that prevents saving filters and settings the first time ([207f0856](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/207f0856)) Closes [#1003](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1003)
+  * Fixed bug in `AbstractOServiceBaseComponent` when searching for a column of type service render ([47a60682](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/47a60682)) Closes[#1010](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1010)
 * **o-email-input**: Fixed bug that incorrectly validated email inputs ([46c6e78](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/46c6e78)) Closes [#999](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/999)
 
 ## 8.7.1 (2022-08-29)